requirements.txt
alembic==1.8.1
anyio==3.6.1
asyncpg==0.26.0
bcrypt==4.0.0
certifi==2022.6.15
cffi==1.15.1
cfgv==3.3.1
click==8.1.3
colorama==0.4.5
cryptography==37.0.4
databases==0.6.1
distlib==0.3.6
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.2.1
fastapi==0.81.0
fastapi-jwt==0.1.8
fastapi-users==10.1.5
fastapi-users-db-sqlalchemy==4.0.3
filelock==3.8.0
greenlet==1.1.3
h11==0.12.0
httpcore==0.15.0
httptools==0.4.0
httpx==0.23.0
httpx-oauth==0.10.1
identify==2.5.3
idna==3.3
makefun==1.14.0
Mako==1.2.2
MarkupSafe==2.1.1
nodeenv==1.7.0
passlib==1.7.4
pipenv==2022.9.2
platformdirs==2.5.2
pre-commit==2.20.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.1
PyJWT==2.4.0
python-decouple==3.6
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.40
starlette==0.19.1
toml==0.10.2
typing_extensions==4.3.0
uvicorn==0.18.3
virtualenv==20.16.4
virtualenv-clone==0.5.7
websockets==10.3
